Items on the stack are now treated as rooted
The stack is by nature acyclic, so anything on the stack does not
need to be collected as long as it stays on the stack. This means we
can use simple reference counting and postpone the garbage collection.

A simple test suggested that this change makes the virtual machine more
than 6% slower.