All kinds of wacky things to ident implementation
Idents are no longer actual Values. They share the first byte (which
for idents is always zero), for type identification purposes, but
that's it. Because idents are only supposed to be created when loading
libraries (such as the standard library) and modules, and don't really
need removed, they can skip garbage collection. Most of the memory used
for Values is just GC housekeeping, which is unnecessary for idents
now. To decrease the number of pointers and fragmentation of heap
memory, ident values are now tied to their ITreeNodes. As a side
effect, idents no longer have hashes based on their contents (which are
unused as of one of the recent commits anyway).