Adds std.vu to be loaded before the main module
It adds three functions:
One called while, which takes two functions, a predicate and a body,
and as long as the predicate function returns true, it runs the body
function. Equivalent to:
func while condition body:
	while condition:
		body

One called $, which takes two functions and returns a function that
calls them both. Basically, function composition. Equivalent to:
$ a b:
	labda:
		a b

The last one is for:, which is also implemented as a macro in the
current compiler. It seemed useful to have it as an actual function, so
it can be passed around and called later. Equivalent to:
for::
	for i:
		i

Example of while and $:

cat:
	while $ @dup @input @print
	drop

Equivalent to:

cat:
	while dup input:
		print
	drop