Changes the way modules are read
They are now read from memory, rather than using file reading
functions. The original read_file function is now a wrapper that first
reads in the whole file, calls the function that actually does all the
work and frees the memory used by the buffer.

This is mainly so I can 'inline' the part of the stdlib implemented in
DVA, so it doesn't have to locate a file called 'std.vu', with
everything that can go wrong:

1) The directory std.vu is in isn't in DEJAVUPATH.
2) There exists another file called std.vu in DEJAVUPATH.
3) There exists a file called std in DEJAVUPATH.
4) Someone deleted std.vu.

There are probably other things that can go wrong. So I'm going to add
a script that generates C code for a function that loads std.vu in
memory. Then, I'm going to replace find_file(get_ident("std")) by a
call to that function. And like magic, we now only have a compile-time
dependency on std.vu.