Adds run_test
It is a little bash script I've been using to compile and
disassemble modules as well as running them. These are
mostly test modules that are in a tests folder, which I
might add in the future as well. Most of those tests are
really simple sanity tests, including one empty module.