Fixes reference clearing bug in exists?
exists? cleared a reference to the object that the ident pointed to,
if present. This caused occasional memory corruption.

I noticed it because after calling exists? :., :. would suddenly
point to :1. If not for that commit that prints the contents of the
stack after the program has returned, I might not have found this.