Fixes bug that only triggered for tiny modules
It doesn't try to read the header of a module and just
gives up if the length of the module is too small anyway.
I compared the length of the module to the size of the
Header struct, which might have been 8 bytes once, but is
now 24 bytes (on my system). So modules with code and
literal data of a total size of less than 16 bytes would be
rejected. This normally never happened because of the debug
information usually included: even null.vu has 28 bytes
after the header. But now I'm testing with persistence,
which doesn't include that debugging information. So I find
this strange thing where :foo and "foo" are rejected and
numbers and :longer-ident and "super long string" are
accepted...