Adds list (makes list from iterator)
This replaces the pattern

    [ for iter ]

by

    list iter

It's currently a little slower, but it might help in those cases where
it is a little hard to sneak a ] before the expression that makes the
iterator (you can't just use swap, because an iterator is 3 values on
the stack).

Functions that currently return a list (*cough* keys *cough*), might
return an iterator in the future, so that

    set :fooList keys foo
    for k in keys foo:
        ...

would be replaced by:

    set :fooList list keys foo
    for k keys foo:
        ...

but this hasn't been decided yet.